requirements.txt
PyPDF2
python-dotenv
langchain
langchain-core
langchain-google-genai
langgraph
faiss-cpu
yfinance
typing-extensions
tiktoken
langchain-community